requirements.txt
streamlit
vaderSentiment
matplotlib
seaborn
google-generativeai
